Fix soft delete and audit Fix mapping value from db to user

+++ src/main/java/com/msm/core/objects/generic/handler/GenericObjectHandler.java
@@ -8,12 +8,13 @@ import com.msm.core.filter.domain.ObjectFilterRequest;
 import com.msm.core.filter.domain.PageResponse;
 import com.msm.core.hook.anontation.Handler;
 import com.msm.core.hook.context.ActionRequest;
+import com.msm.core.metadata.Attribute;
 import com.msm.core.metadata.ObjectMetadata;
 import com.msm.core.objects.exception.Errors;
 import com.msm.core.objects.exception.ServiceErrorEnum;
 import com.msm.core.objects.generic.audit.AuditAction;
 import com.msm.core.objects.generic.audit.AuditStrategy;
-import com.msm.core.objects.generic.converter.ObjectMappingStrategy;
+import com.msm.core.objects.generic.converter.CustomValueMappingStrategy;
 import com.msm.core.objects.generic.service.GenericObjectMetadataService;
 import com.msm.core.strategy.StrategyResolver;
 import lombok.RequiredArgsConstructor;
@@ -27,38 +28,37 @@ public class GenericObjectHandler {
     private final DynamicQueryService dynamicQueryService;
     private final GenericObjectMetadataService genericObjectMetadataService;
     private final StrategyResolver<AuditStrategy> auditStrategyFactory;
-    private final StrategyResolver<ObjectMappingStrategy> objectMappingStrategyFactory;
+    private final StrategyResolver<CustomValueMappingStrategy> objectMappingStrategyFactory;
 
     @Handler(action = Constants.GENERIC_FILTER_ACTION)
-    public <X> X filter(ActionRequest<ObjectFilterRequest> request) {
+    public PageResponse<Map<String, Object>> filter(ActionRequest<ObjectFilterRequest> request) {
         ObjectMetadata objectMetadata = getObjectMetadata(request.getObjectName());
-        PageResponse<Object> pageResponse = dynamicQueryService.filter(objectMetadata, request.getPayload());
-        mapFrom(objectMetadata, pageResponse.getContents());
-        return (X) pageResponse;
+        PageResponse<Map<String, Object>> pageResponse = dynamicQueryService.filter(objectMetadata, request.getPayload());
+        Utils.CL.emptyIfNull(pageResponse.getContents()).forEach(object -> mapFrom(objectMetadata, object));
+        return pageResponse;
     }
 
     @Handler(action = Constants.GENERIC_ALL_OBJECT_ACTION)
-    public <X> X findAllObject(ActionRequest<ObjectFilterRequest> request) {
+    public List<Map<String, Object>> findAllObject(ActionRequest<ObjectFilterRequest> request) {
         ObjectMetadata objectMetadata = getObjectMetadata(request.getObjectName());
-        PageResponse<Object> pageResponse = dynamicQueryService.filter(objectMetadata, request.getPayload());
-        mapFrom(objectMetadata, pageResponse.getContents());
-        return (X) pageResponse.getContents();
+        PageResponse<Map<String, Object>> pageResponse = dynamicQueryService.filter(objectMetadata, request.getPayload());
+        Utils.CL.emptyIfNull(pageResponse.getContents()).forEach(object -> mapFrom(objectMetadata, object));
+        return pageResponse.getContents();
     }
 
     @Handler(action = Constants.GENERIC_FILTER_BY_ID_ACTION)
     public <X> X findObjectById(ActionRequest<ObjectFilterRequest> request) {
         ObjectMetadata objectMetadata = getObjectMetadata(request.getObjectName());
         PageResponse<Object> pageResponse = dynamicQueryService.filter(objectMetadata, request.getPayload());
-        List<Object> objects = pageResponse.getContents();
-        mapFrom(objectMetadata, objects);
-        return (X) Utils.CL.getFirst(objects);
+        Utils.CL.emptyIfNull(pageResponse.getContents()).forEach(object -> mapFrom(objectMetadata, (Map<String, Object>) object));
+        return (X) Utils.CL.getFirst(pageResponse.getContents());
     }
 
     @Handler(action = Constants.Action.CREATE)
     public <X> X create(ActionRequest<Map<String, Object>> request) {
         ObjectMetadata objectMetadata = getObjectMetadata(request.getObjectName());
-        mapTo(objectMetadata, request.getPayload());
         applyAudit(objectMetadata, AuditAction.CREATE, request.getPayload());
+        mapTo(objectMetadata, request.getPayload());
         Map<String, Object> returnObject = dynamicQueryService.insertReturning(objectMetadata, request.getPayload());
         mapFrom(objectMetadata, returnObject);
         return (X) returnObject;
@@ -78,31 +78,27 @@ public class GenericObjectHandler {
         } catch (JsonMappingException e) {
             throw Errors.throwException(ServiceErrorEnum.INVALID_UPDATE_PAYLOAD, e.getOriginalMessage());
         }
-        mapTo(objectMetadata, oldData);
         applyAudit(objectMetadata, AuditAction.UPDATE, oldData);
+        mapTo(objectMetadata, oldData);
         dynamicQueryService.updateById(objectMetadata, request.getObjectId(), oldData);
         mapFrom(objectMetadata, oldData);
         return (X) oldData;
     }
 
     @Handler(action = Constants.Action.DELETE)
-    public <X> X delete(ActionRequest<Map<String, Object>> request) {
+    public int delete(ActionRequest<Map<String, Object>> request) {
         ObjectMetadata objectMetadata = getObjectMetadata(request.getObjectName());
+        if(!isSoftDeleted(objectMetadata)) {
+            return dynamicQueryService.forceDeleteById(objectMetadata, request.getObjectId());
+        }
         applyAudit(objectMetadata, AuditAction.DELETE, request.getPayload());
-        Object rowEffected = dynamicQueryService.deleteById(objectMetadata, request.getObjectId(), request.getPayload());
-        return (X) rowEffected;
+        return dynamicQueryService.deleteById(objectMetadata, request.getObjectId(), request.getPayload());
     }
 
     private ObjectMetadata getObjectMetadata(String objectName) {
         return genericObjectMetadataService
                 .getObjectMetadata(objectName)
                 .orElseThrow(() -> Errors.throwException(ServiceErrorEnum.OBJECT_META_DATA_NOT_FOUND));
-    }
-
-    private void mapFrom(ObjectMetadata objectMetadata, List<Object> objects) {
-        Utils.CL.emptyIfNull(objects).forEach(o -> {
-            mapFrom(objectMetadata, (Map<String, Object>) o);
-        });
     }
 
     private void mapFrom(ObjectMetadata objectMetadata, Map<String, Object> object) {
@@ -121,5 +117,13 @@ public class GenericObjectHandler {
         auditStrategyFactory
                 .resolve(objectMetadata.getName())
                 .apply(action, objectMetadata, data);
+    }
+
+    private boolean isSoftDeleted(ObjectMetadata meta) {
+        Attribute deletedAt = meta.getAttributeByName(Constants.DELETED_AT);
+        Attribute deletedBy = meta.getAttributeByName(Constants.DELETED_BY);
+        Attribute deletedById = meta.getAttributeByName(Constants.DELETED_BY_ID);
+        Attribute isDeleted = meta.getAttributeByName(Constants.IS_DELETED);
+        return isDeleted != null || deletedAt != null || deletedBy != null || deletedById != null;
     }
 }
